pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "openpectus_database_administration"
authors = [
  {name = "Eskild Schroll-Fleischer", email = "eyfl@novonordisk.com"},
]
dynamic = ["version", "description"]
readme = "README.md"
requires-python = ">= 3.11"

dependencies = [
    #"openpectus --no-deps",
    "pydantic==2.10.4",
    "uvicorn[standard]==0.34.0",
    "sqlalchemy==2.0.36",
    "starlette",
    "itsdangerous",
    "starlette-admin",
    "msal",
    "PyJWT",
    "uvicorn",
]

keywords = ["database administration"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Natural Language :: English",
  "Operating System :: Microsoft :: Windows",
  "Operating System :: POSIX :: Linux",
  "Programming Language :: Python",
  "Topic :: System",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]

[project.optional-dependencies]
development = [
    "pyright==1.1.400",
    "flake8==7.1.1",
    "black==24.10.0",
    "build==1.2.2.post1",
]

[project.urls]
Documentation = "https://docs.openpectus.org/"
Repository = "https://github.com/Open-Pectus/Database-Administration.git"
GitHub = "https://github.com/Open-Pectus/Database-Administration"
Issues = "https://github.com/Open-Pectus/Database-Administration/issues"

[tool.flit.sdist]
include = ["openpectus_database_administration"]

[tool.flit.module]
name = "openpectus_database_administration"

[project.scripts]
pectus-database-administration = "openpectus_database_administration.app:main"

[tool.pyright]
include = ["openpectus_database_administration"]
exclude = [
    "**/__pycache__",
]
